AnnotatedCFG: last_statement_index() returns None if the block has exits.
In SimEngineVEX, if last statement ID is less than or equal to the last
statement of a block, the default exit will not be executed. Therefore
if the default exit of a basic block should be executed, the last
statement ID must be a number greater than the total number of
statements of a block.